1. Se agregan los margenes de contenido en todos los jsp del módulo de seguridad 2. Se corrige: "El módulo de usuarios está permitiendo eliminar el usuario admin, el problema aquí es que cualquier usuario puede borrar todos los usuarios." Solución: Ya no se permite el borrado de usuarios pertenecientes al grupo de administradores. 3. Se implementa: "Al eliminar un usuario de sistema se debe preguntar al usuario si realmente está seguro de realizar dicha acción." 4. Se corrige: "Acomodar los componentes gráficos en la pantalla de crear nuevo usuario."
git-svn-id: svn://pulsatrix.inbio.ac.cr/atta2/trunk@1558 3aa2e761-397d-4522-80f4-092673de1051

+++ ara-web/src/java/org/inbio/ara/security/ListUser.java
@@ -23,10 +23,12 @@ import java.util.ArrayList;
 import java.util.Locale;
 import javax.faces.FacesException;
 import javax.faces.component.html.HtmlDataTable;
+import javax.faces.component.html.HtmlInputHidden;
 import javax.faces.context.FacesContext;
 import org.inbio.ara.AraSessionBean;
 import org.inbio.ara.dto.security.SystemUserDTO;
 import org.inbio.ara.util.AddRemoveList;
+import org.inbio.ara.util.BundleHelper;
 import org.inbio.ara.util.MessageBean;
 
 
@@ -61,6 +63,7 @@ public class ListUser extends AbstractPageBean {
 
     //Bindings de componentes graficos
     private HtmlDataTable dataTableUser = new HtmlDataTable();
+    private HtmlInputHidden deleteConfirmationText = new HtmlInputHidden();
 
     //Variable que contiene los datos de la paginacion para ser mostrados en la tabla
     private String quantityTotal = new String();
@@ -128,6 +131,10 @@ public class ListUser extends AbstractPageBean {
      */
     @Override
     public void prerender() {
+        //Setting the confirmation text for deleting option
+        this.deleteConfirmationText.setValue(BundleHelper.getDefaultBundleValue
+                    ("delete_confirmation", this.getMyLocale()));
+        //Manage pagination
         if(this.getUserSessionBean().getPagination()==null){
             this.getUserSessionBean().initDataProvider();
         }
@@ -244,10 +251,16 @@ public class ListUser extends AbstractPageBean {
         else if(selectedUsers.size() == 1){ //En caso de que solo se seleccione un elemento
 
             SystemUserDTO selected = selectedUsers.get(0);
-            //Mandar a eliminar las entradas en user_taxon
-            this.getUserSessionBean().deleteUserTaxonsByUser(selected.getUserId());
 
             try{
+                //Verify if user is administrator
+                if(this.getUserSessionBean().isAdmin(selected.getUserId())){
+                    MessageBean.setErrorMessageFromBundle("imposible_to_delete", this.getMyLocale());
+                    return null;
+                }
+
+                //Mandar a eliminar las entradas en user_taxon
+                this.getUserSessionBean().deleteUserTaxonsByUser(selected.getUserId());
                 //Mandar a eliminar las entradas de user_nomenclatural_group
                 this.getUserSessionBean().deleteNomenclaturalGroupsByUser(selected.getUserId());
                 //Mandar a eliminar la entidad correspondiente al dto
@@ -255,6 +268,7 @@ public class ListUser extends AbstractPageBean {
             }
             catch(Exception e){
                 MessageBean.setErrorMessageFromBundle("imposible_to_delete", this.getMyLocale());
+                System.out.print("mata"+e);
                 return null;
             }
             
@@ -264,7 +278,6 @@ public class ListUser extends AbstractPageBean {
 
             //Notificar al usuario
             MessageBean.setSuccessMessageFromBundle("delete_user_success", this.getMyLocale());
-
             return null;
 
         }
@@ -328,6 +341,20 @@ public class ListUser extends AbstractPageBean {
      */
     public Locale getMyLocale() {
 		return this.getAraSessionBean().getCurrentLocale();
+    }
+
+    /**
+     * @return the deleteConfirmationText
+     */
+    public HtmlInputHidden getDeleteConfirmationText() {
+        return deleteConfirmationText;
+    }
+
+    /**
+     * @param deleteConfirmationText the deleteConfirmationText to set
+     */
+    public void setDeleteConfirmationText(HtmlInputHidden deleteConfirmationText) {
+        this.deleteConfirmationText = deleteConfirmationText;
     }
 
 }
